+++ src/formats.rs
@@ -6,7 +6,8 @@
 //! See [the specification](https://github.com/GFA-spec/GFA-spec/blob/master/GFA1.md) for details.
 //! The following functions support line-by-line writing of GFA version 1.1:
 //!
-//! * [`write_gfa_header`]: Write the header line.
+//! * [`write_gfa_header`]: Write a GFA file header.
+//! * [`write_header_lines`]: Write additional header lines.
 //! * [`write_gfa_segment`]: Write a segment line for a node.
 //! * [`write_gfa_link`]: Write a link line for an edge.
 //! * [`write_gfa_walk`]: Write a walk line for a path.
@@ -26,13 +27,20 @@
 //! See [the specification](https://github.com/lh3/gfatools/blob/master/doc/rGFA.md) for an overview.
 //! Some details are better documented in the [minimap2 man page](https://lh3.github.io/minimap2/minimap2.html#10).
 //!
-//! GAF I/O is currently implemented in [`crate::Alignment`].
-//! [`is_gaf_header_line`] can be used to identify header lines when reading GAF files.
+//! GAF header lines can be handled with the following functions:
+//!
+//! * [`is_gaf_header_line`]: Check if a buffer contains a GAF header line.
+//! * [`peek_gaf_header_line`]: Check if the next line in a reader would be a GAF header line.
+//! * [`read_gaf_header_lines`]: Read all successive GAF header lines from a reader.
+//! * [`write_gaf_file_header`]: Write a GAF file header.
+//! * [`write_header_lines`]: Write additional header lines.
+//!
+//! I/O for GAF alignment lines is currently implemented in [`crate::Alignment`].
 
 use crate::utils;
 
 use std::fmt::Display;
-use std::io::{self, Write};
+use std::io::{self, BufRead, Write};
 use std::ops::Range;
 use std::str;
 
@@ -272,13 +280,64 @@ impl WalkMetadata {
 
 //-----------------------------------------------------------------------------
 
-// TODO: Tests when we start parsing GAF headers.
-
 /// Returns `true` if the buffer contains a GAF header line.
 ///
 /// A header line starts with `@`.
 pub fn is_gaf_header_line(buf: &[u8]) -> bool {
     buf.first() == Some(&b'@')
+}
+
+/// Returns `true` if the next line in the reader would be a GAF header line.
+///
+/// The reader position is not changed.
+/// Returns an I/O error if reading from the reader fails.
+/// A header line starts with `@`.
+pub fn peek_gaf_header_line<R: BufRead>(reader: &mut R) -> io::Result<bool> {
+    let buffer = reader.fill_buf()?;
+    Ok(buffer.first() == Some(&b'@'))
+}
+
+/// Returns all successive GAF header lines from the reader.
+///
+/// The returned lines do not contain the trailing newline character.
+/// The reader position is advanced past the header lines.
+/// Returns an I/O error if reading from the reader fails.
+///
+/// # Examples
+///
+/// ```
+/// use gbz_base::{formats, utils};
+///
+/// let filename = utils::get_test_data("good.gaf");
+/// let mut reader = utils::open_file(&filename)
+///     .expect("Unable to open test file");
+/// let headers = formats::read_gaf_header_lines(&mut reader)
+///     .expect("Unable to read GAF header lines");
+/// assert_eq!(headers.len(), 2);
+/// assert!(headers[0].starts_with("@HD")); // File header
+/// assert!(headers[1].starts_with("@RN")); // Reference name
+/// ```
+pub fn read_gaf_header_lines<R: BufRead>(reader: &mut R) -> io::Result<Vec<String>> {
+    let mut headers: Vec<String> = Vec::new();
+    while peek_gaf_header_line(reader)? {
+        let mut line: Vec<u8> = Vec::new();
+        let bytes_read = reader.read_until(b'\n', &mut line)?;
+        if bytes_read == 0 {
+            break;
+        }
+        if line.last() == Some(&b'\n') {
+            line.pop();
+        }
+        headers.push(String::from_utf8_lossy(&line).to_string());
+    }
+    Ok(headers)
+}
+
+/// Writes a GAF file header.
+pub fn write_gaf_file_header<T: Write>(output: &mut T) -> io::Result<()> {
+    let header = format!("@HD\tVN:Z:1.0\n");
+    output.write_all(header.as_bytes())?;
+    Ok(())
 }
 
 //-----------------------------------------------------------------------------
@@ -297,6 +356,21 @@ pub fn write_gfa_header<T: Write>(reference_samples: Option<&str>, output: &mut 
         "H\tVN:Z:1.1\n".to_string()
     };
     output.write_all(header.as_bytes())?;
+    Ok(())
+}
+
+/// Writes the given header lines.
+///
+/// If the lines do not end with a newline character, one is added.
+pub fn write_header_lines<T: Write>(header_lines: &[String], output: &mut T) -> io::Result<()> {
+    for line in header_lines {
+        if line.ends_with('\n') {
+            output.write_all(line.as_bytes())?;
+        } else {
+            output.write_all(line.as_bytes())?;
+            output.write_all(b"\n")?;
+        }
+    }
     Ok(())
 }