+++ RELEASES.md
@@ -3,6 +3,10 @@
 ## Current version
 
 * `db2gaf` tool for converting a GAF-base back to GAF format.
+* Support for [stable graph names](https://github.com/jltsiren/pggname):
+  * Uses `pggname::GraphName` for importing graph names and relationships between GBZ tags and GFA/GAF headers.
+  * Stable graph names are stored in databases and included in GFA/GAF outputs.
+  * `db2gaf` and `query` use the information for determining if the graph is a valid reference for the alignments.
 
 ## GBZ-base 0.1.0 (2025-11-06)